refactor(http): split Node server bootstrap from Hono app builder (#124)
The Worker bundle only consumes `createHttpApp` from `httpTransport.ts`, but
the file also imported `node:http` and `@hono/node-server`'s `serve()` for
the Node-process bootstrap path. Both modules are referenced unconditionally
at the top of the module and survive tree-shaking, bloating the Worker
bundle and hitting `nodejs_compat` polyfill gaps at evaluation time.

Move `startHttpTransport`, `stopHttpTransport`, `HttpTransportHandle`, and
the `node:http` port probe to a new `httpServer.ts`. `manager.ts` is the
only consumer of those exports and re-points; the lifecycle and manager
tests re-point as well. `httpTransport.ts` now exports the runtime-agnostic
`createHttpApp` only.
